+++ lib/manager/index.js
@@ -22,16 +22,25 @@ const get = (manager, name) => managers[manager][name];
 const getLanguageList = () => languageList;
 const getManagerList = () => managerList;
 
-function postExtract(manager, packageFiles) {
-  if (managers[manager].postExtract) {
-    return managers[manager].postExtract(packageFiles);
-  }
-  return null;
-}
-
 module.exports = {
   get,
   getLanguageList,
   getManagerList,
-  postExtract,
 };
+
+const managerFunctions = [
+  'extractDependencies',
+  'postExtract',
+  'getPackageUpdates',
+  'updateDependency',
+  'supportsLockFileMaintenance',
+];
+
+for (const f of managerFunctions) {
+  module.exports[f] = (manager, ...params) => {
+    if (managers[manager][f]) {
+      return managers[manager][f](...params);
+    }
+    return null;
+  };
+}
