+++ lib/manager/npm/extract/index.js
@@ -9,6 +9,7 @@ module.exports = {
 };
 
 async function extractDependencies(content, packageFile) {
+  logger.debug({ content, packageFile });
   const deps = [];
   let packageJson;
   try {
