+++ lib/workers/repository/extract/index.js
@@ -1,5 +1,5 @@
-const { getManagerList } = require('../../manager');
-const { getManagerConfig } = require('../../config');
+const { getManagerList } = require('../../../manager');
+const { getManagerConfig } = require('../../../config');
 const { getManagerPackageFiles } = require('./manager-files');
 
 module.exports = {
