+++ lib/workers/repository/extract/manager-files.js
@@ -2,7 +2,7 @@ module.exports = {
   getManagerPackageFiles,
 };
 
-const { extractDependencies, postExtract } = require('../../manager');
+const { extractDependencies, postExtract } = require('../../../manager');
 
 const {
   getIncludedFiles,
