+++ lib/workers/repository/index.js
@@ -13,17 +13,13 @@ async function renovateRepository(repoConfig) {
   let config = { ...repoConfig };
   logger.setMeta({ repository: config.repository });
   logger.info('Renovating repository');
-  logger.trace({ config }, 'renovateRepository()');
+  logger.trace({ config });
   try {
     config = await initRepo(config);
-    let res;
-    let branches;
-    let branchList;
-    let packageFiles;
-    ({ res, branches, branchList, packageFiles } = await processRepo(config)); // eslint-disable-line prefer-const
-    if (!config.repoIsOnboarded) {
-      res = await ensureOnboardingPr(config, packageFiles, branches);
-    }
+    const { res, branches, branchList, packageFiles } = await processRepo(
+      config
+    );
+    await ensureOnboardingPr(config, packageFiles, branches);
     await finaliseRepo(config, branchList);
     return processResult(config, res);
   } catch (err) /* istanbul ignore next */ {
