+++ lib/workers/repository/onboarding/pr/index.js
@@ -4,6 +4,9 @@ const { getBaseBranchDesc } = require('./base-branch');
 const { getPrList } = require('./pr-list');
 
 async function ensureOnboardingPr(config, packageFiles, branches) {
+  if (config.repoIsOnboarded) {
+    return;
+  }
   logger.debug('ensureOnboardingPr()');
   logger.trace({ config });
   const onboardingBranch = `renovate/configure`;
@@ -65,12 +68,12 @@ You can post questions in [our Config Help repository](https://github.com/renova
     // Check if existing PR needs updating
     if (existingPr.title === onboardingPrTitle && existingPr.body === prBody) {
       logger.info(`${existingPr.displayNumber} does not need updating`);
-      return 'onboarding';
+      return;
     }
     // PR must need updating
     await platform.updatePr(existingPr.number, onboardingPrTitle, prBody);
     logger.info(`Updated ${existingPr.displayNumber}`);
-    return 'onboarding';
+    return;
   }
   logger.info('Creating onboarding PR');
   const labels = [];
@@ -83,7 +86,6 @@ You can post questions in [our Config Help repository](https://github.com/renova
     useDefaultBranch
   );
   logger.info({ pr: pr.displayNumber }, 'Created onboarding PR');
-  return 'onboarding';
 }
 
 module.exports = {
