+++ lib/workers/repository/process/extract-update.js
@@ -1,15 +1,15 @@
 const { writeUpdates } = require('./write');
 const { sortBranches } = require('./sort');
-const { get } = require('../../../manager');
 const { fetchUpdates } = require('./fetch');
 const { branchifyUpgrades } = require('../updates/branchify');
+const { extractDependencies } = require('../extract');
 
 module.exports = {
   extractAndUpdate,
 };
 
 async function extractAndUpdate(config) {
-  const extractDependencies = get(config.manager, 'extractDependencies');
+  logger.debug('extractAndUpdate()');
   const packageFiles = await extractDependencies(config);
   logger.debug({ packageFiles }, 'packageFiles');
   await fetchUpdates(config, packageFiles);
