+++ lib/workers/repository/process/index.js
@@ -6,6 +6,7 @@ module.exports = {
 };
 
 async function processRepo(config) {
+  logger.debug('processRepo()');
   if (config.baseBranches && config.baseBranches.length) {
     logger.info({ baseBranches: config.baseBranches }, 'baseBranches');
     let res;
@@ -24,5 +25,6 @@ async function processRepo(config) {
     }
     return { res, branches, branchList };
   }
+  logger.debug('No baseBranches');
   return extractAndUpdate(config);
 }
