+++ lib/workers/repository/result.js
@@ -31,11 +31,7 @@ function processResult(config, result) {
     status = 'enabled';
   } else {
     status = 'onboarding';
-    if (result === 'onboarding') {
-      res = 'done';
-    } else {
-      res = result;
-    }
+    res = 'done';
   }
   return { res, status };
 }
